Remove Query from Requirement directory name
