+++ src/RepoAuditor/Plugins/GitHub/StandardRequirements/DefaultBranch.py
@@ -8,7 +8,7 @@
 
 import textwrap
 
-from RepoAuditor.Plugins.GitHub.StandardQueryRequirements.Impl.StandardValueRequirementImpl import (
+from RepoAuditor.Plugins.GitHub.StandardRequirements.Impl.StandardValueRequirementImpl import (
     StandardValueRequirementImpl,
 )
