+++ src/RepoAuditor/Plugins/GitHub/StandardRequirements/MergeCommit.py
@@ -4,42 +4,40 @@
 # |  Distributed under the MIT License.
 # |
 # -------------------------------------------------------------------------------
-"""Contains the RebaseMergeCommit object."""
+"""Contains the MergeCommit object."""
 
 import textwrap
 
-from RepoAuditor.Plugins.GitHub.StandardQueryRequirements.Impl.StandardEnableRequirementImpl import (
+from RepoAuditor.Plugins.GitHub.StandardRequirements.Impl.StandardEnableRequirementImpl import (
     StandardEnableRequirementImpl,
 )
 
 
 # ----------------------------------------------------------------------
-class RebaseMergeCommit(StandardEnableRequirementImpl):
-    """Allow rebase merging requirement."""
+class MergeCommit(StandardEnableRequirementImpl):
+    """Allow merge commits requirement."""
 
     # ----------------------------------------------------------------------
     def __init__(self) -> None:
         super().__init__(
-            "RebaseMergeCommit",
-            False,
-            "true",
+            "MergeCommit",
+            True,
+            "false",
             "settings",
             "Pull Requests",
-            "Allow rebase merging",
-            lambda data: data["standard"].get("allow_rebase_merge", None),
+            "Allow merge commits",
+            lambda data: data["standard"].get("allow_merge_commit", None),
             textwrap.dedent(
                 """\
-                The default behavior is to not allow rebase merging.
+                The default behavior is to allow merge commits.
 
                 Reasons for this Default
                 ------------------------
-                - Rebase merging is not compatible with signed commits, as GitHub creates a new commit when rebasing.
+                - Merge commits are the most basic way to merge from a branch into another branch.
 
                 Reasons to Override this Default
                 --------------------------------
-                - Your repository does not require signatures.
-                - You want GitHub to rebase for you as part of the pull request process when changes by others are frequent or the
-                  pull request process can last for an extended period of time.
+                <unknown>
                 """,
             ),
         )
