+++ src/RepoAuditor/Plugins/GitHub/StandardRequirements/MergeCommitMessage.py
@@ -4,50 +4,50 @@
 # |  Distributed under the MIT License.
 # |
 # -------------------------------------------------------------------------------
-"""Contains the SquashMergeCommitMessage object."""
+"""Contains the MergeCommitMessage object."""
 
 import textwrap
 from typing import Any
 
-from RepoAuditor.Plugins.GitHub.StandardQueryRequirements.Impl.StandardValueRequirementImpl import (
+from RepoAuditor.Plugins.GitHub.StandardRequirements.Impl.StandardValueRequirementImpl import (
     DoesNotApplyResult,
     StandardValueRequirementImpl,
 )
 
 
 # ----------------------------------------------------------------------
-class SquashMergeCommitMessage(StandardValueRequirementImpl):
-    """Set default commit message for squash merges."""
+class MergeCommitMessage(StandardValueRequirementImpl):
+    """Allow merge commit message requirement."""
 
     # ----------------------------------------------------------------------
     def __init__(self) -> None:
         super().__init__(
-            "SquashMergeCommitMessage",
-            "COMMIT_MESSAGES",
+            "MergeCommitMessage",
+            "BLANK",
             "settings",
             "Pull Requests",
-            "Allow squash merging -> Default...",
+            "Allow merge commits -> Default...",
             _GetValue,
             textwrap.dedent(
                 """\
                 Available values:
 
+                PR_TITLE [Default message]
+                    Pull Request number and head branch on the first line; pull request title on the third line
+
                 BLANK [Default to pull request title]
                     Pull Request title and number on the first line.
 
-                COMMIT_MESSAGES [Default to pull request title and commit details]
-                    Commit title and...
-                        [Single Commit] ...commit message
-                        [Multiple Commits] ...pull request title and number and list of commits
-
                 PR_BODY [Default to pull request title and description]
-                    Pull Request title and number on the first line; commit description starting on the third line.
+                    Pull Request title and number on the first line; pull request description starting on the third line.
 
-            The default setting is COMMIT_MESSAGES.
+            The default setting is BLANK.
 
             Reasons for this Default
             ------------------------
-            - Preserves the information in the original commits.
+            - Reduce redundant information by only duplicating the title of the commit(s).
+            - PR_TITLE includes the head branch name, which oftentimes is not relevant information to preserve over time.
+            - PR_BODY duplicates the title and description of the commit(s).
 
             Reasons to Override this Default
             --------------------------------
@@ -63,11 +63,11 @@ class SquashMergeCommitMessage(StandardValueRequirementImpl):
 def _GetValue(data: dict[str, Any]) -> str | DoesNotApplyResult | None:
     data = data["standard"]
 
-    result = data.get("allow_squash_merge", None)
+    result = data.get("allow_merge_commit", None)
     if result is None:
         return None
 
     if not result:
-        return DoesNotApplyResult("Squash merge commits are not enabled.")
+        return DoesNotApplyResult("Merge commits are not enabled.")
 
-    return data.get("squash_merge_commit_message", None)
+    return data.get("merge_commit_message", None)
