+++ src/RepoAuditor/Plugins/GitHub/StandardRequirements/SecretScanningPushProtection.py
@@ -4,42 +4,41 @@
 # |  Distributed under the MIT License.
 # |
 # -------------------------------------------------------------------------------
-"""Contains the DependabotSecurityUpdates object."""
+"""Contains the SecretScanningPushProtection object."""
 
 import textwrap
 from typing import Any, Optional
 
-from RepoAuditor.Plugins.GitHub.StandardQueryRequirements.Impl.StandardEnableRequirementImpl import (
+from RepoAuditor.Plugins.GitHub.StandardRequirements.Impl.StandardEnableRequirementImpl import (
     StandardEnableRequirementImpl,
 )
 
 
 # ----------------------------------------------------------------------
-class DependabotSecurityUpdates(StandardEnableRequirementImpl):
-    """Requirement of Dependabot security updates."""
+class SecretScanningPushProtection(StandardEnableRequirementImpl):
+    """Push protection for secret scanning."""
 
     # ----------------------------------------------------------------------
     def __init__(self) -> None:
         super().__init__(
-            "DependabotSecurityUpdates",
+            "SecretScanningPushProtection",
             True,
             "disabled",
             "settings/security_analysis",
-            "Dependabot",
-            "Dependabot security updates",
+            "Secret Protection",
+            "Push protection",
             _GetValue,
             textwrap.dedent(
                 """\
-                The default behavior is to ensure Dependabot security updates are enabled.
+                The default behavior is to ensure secret scanning push protection is enabled.
 
                 Reasons for this Default
                 ------------------------
-                - Increases the overall security of the repository by automatically applying security updates.
+                - Secrets should not be checked into code.
 
                 Reasons to Override this Default
                 --------------------------------
-                - Dependabot security updates are not supported for the repository or by the organization.
-                - A manual test pass is required before changes can be deployed.
+                <unknown>
                 """,
             ),
             unset_set_terminology=("disabled", "enabled"),
@@ -53,7 +52,7 @@ def _GetValue(data: dict[str, Any]) -> Optional[bool]:
     result = (
         data["standard"]
         .get("security_and_analysis", {})
-        .get("dependabot_security_updates", {})
+        .get("secret_scanning_push_protection", {})
         .get("status", None)
     )
